+++ discovery/moby/tasks_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dockerswarm
+package moby
 
 import (
 	"context"
@@ -35,7 +35,7 @@ func TestDockerSwarmTasksSDRefresh(t *testing.T) {
 role: tasks
 host: %s
 `, url)
-	var cfg SDConfig
+	var cfg DockerSwarmSDConfig
 	require.NoError(t, yaml.Unmarshal([]byte(cfgString), &cfg))
 
 	d, err := NewDiscovery(&cfg, log.NewNopLogger())
