+++ discovery/moby/testdata/dockerprom/_ping.json
