+++ discovery/moby/testdata/dockerprom/networks.json
