+++ discovery/moby/testdata/swarmprom/_ping.json
