+++ discovery/moby/testdata/swarmprom/headers.yml
