+++ discovery/moby/testdata/swarmprom/networks.json
