+++ discovery/moby/testdata/swarmprom/nodes.json
