+++ discovery/moby/testdata/swarmprom/services.json
