+++ discovery/moby/testdata/swarmprom/services.json__wvKVCQ4HhD
