+++ discovery/moby/testdata/swarmprom/tasks.json
